Fork old tasks graph and implement feature flag
