+++ ydb/core/kqp/executer_actor/tasks_graph/kqp_tasks_graph.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "max_tasks_graph.h"
 #include "shard_key_ranges.h"
 
 #include <ydb/core/kqp/common/kqp_user_request_context.h>
@@ -22,6 +21,7 @@ namespace NKikimr::NKqp {
 class TPartitionPruner;
 struct TPartitionPrunerConfig;
 struct TQueryExecutionStats;
+struct TPhysicalShardReadSettings;
 
 struct TTransaction : private TMoveOnly {
     NYql::NNodes::TKqpPhysicalTx Node;
@@ -398,11 +398,11 @@ public:
         const TVector<IKqpGateway::TPhysicalTxData>& transactions,
         const NKikimr::NKqp::TTxAllocatorState::TPtr& txAlloc,
         const NKikimrConfig::TTableServiceConfig::TAggregationConfig& aggregationSettings,
-        const NKikimrConfig::TTableServiceConfig::TResourceManager& channelSettings,
         const TKqpRequestCounters::TPtr& counters,
         TActorId bufferActorId,
         TIntrusiveConstPtr<NACLib::TUserToken> userToken
     );
+    virtual ~TKqpTasksGraph() = default;
 
     void ResolveShards(TGraphMeta::TShardToNodeMap&& shardsToNodes);
 
@@ -421,26 +421,35 @@ public:
     TVector<TString> GetStageIntrospection(const NYql::NDq::TStageId& stageId) const;
     TString DumpToString() const;
 
-private:
-    void FillStages();
+protected:
+    struct TStageScheduleInfo {
+        double StageCost = 0.0;
+        ui32 TaskCount = 0;
+    };
 
-    void CountScanTasksFromSource(const TStageInfo& stageInfo, bool limitTasksPerNode);
-    void CountFullTextScanTasksFromSource(const TStageInfo& stageInfo);
-    void CountSysViewTasksFromSource(const TStageInfo& stageInfo);
-    void CountReadTasksFromSource(const TStageInfo& stageInfo, size_t resourceSnapshotSize, ui32 scheduledTaskCount);
-    void CountSysViewScanTasks(const TStageInfo& stageInfo);
-    void CountComputeTasks(const TStageInfo& stageInfo, const ui32 nodesCount);
-    void CountScanTasksFromShards(const TStageInfo& stageInfo, bool enableShuffleElimination);
+    // TODO: here only for tasks graph restoration.
+    static std::map<ui32, TStageScheduleInfo> ScheduleByCost(const IKqpGateway::TPhysicalTxData& tx, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot);
+    static void FillReadTaskFromSource(TTask& task, const TString& sourceName, const TString& structuredToken, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot, ui64 nodeOffset);
 
-    void BuildSysViewScanTasks(TStageInfo& stageInfo);
-    void BuildComputeTasks(TStageInfo& stageInfo); // returns true if affected shards count is unknown
-    void BuildDatashardTasks(TStageInfo& stageInfo, THashSet<ui64>* shardsWithEffects); // returns shards with effects
-    void BuildScanTasksFromShards(TStageInfo& stageInfo, bool enableShuffleElimination, TQueryExecutionStats* stats);
-    void BuildFullTextScanTasksFromSource(TStageInfo& stageInfo, TQueryExecutionStats* stats);
-    void BuildSysViewTasksFromSource(TStageInfo& stageInfo);
-    void BuildReadTasksFromSource(TStageInfo& stageInfo, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot, ui32 scheduledTaskCount);
+    // TODO: remove later with the old tasks graph.
+    static void MergeReadInfoToTaskMeta(TTaskMeta& meta, ui64 shardId, TMaybe<TShardKeyRanges>& keyReadRanges,
+        const TPhysicalShardReadSettings& readSettings, const TVector<TTaskMeta::TColumn>& columns,
+        const NKqpProto::TKqpPhyTableOperation& op, bool isPersistentScan);
+    static void PrepareScanMetaForUsage(TTaskMeta& meta, const TVector<NScheme::TTypeInfo>& keyTypes);
+    static TVector<TTaskMeta::TColumn> BuildKqpColumns(const NKqpProto::TKqpReadRangesSource& op, TIntrusiveConstPtr<TTableConstInfo> tableInfo);
+    static TVector<TTaskMeta::TColumn> BuildKqpColumns(const NKqpProto::TKqpPhyTableOperation& op, TIntrusiveConstPtr<TTableConstInfo> tableInfo);
+
     void FillScanTaskLockTxId(NKikimrTxDataShard::TKqpReadRangesSourceSettings& settings);
-    TMaybe<size_t> BuildScanTasksFromSource(TStageInfo& stageInfo, TQueryExecutionStats* stats);
+    std::pair<ui32, TKqpTasksGraph::TTaskType::ECreateReason> GetMaxTasksAggregation(const TStageInfo& stageInfo, const ui32 previousTasksCount, const ui32 nodesCount) const;
+    std::pair<ui32, TKqpTasksGraph::TTaskType::ECreateReason> GetScanTasksPerNode(const TStageInfo& stageInfo, const bool isOlapScan, const ui64 nodeId, bool enableShuffleElimination = false) const;
+    TString ReplaceStructuredTokenReferences(const TString& token) const;
+
+private:
+    virtual size_t DoBuildAllTasks(const TVector<NKikimrKqp::TKqpNodeResources>& resourcesSnapshot, TQueryExecutionStats* stats) = 0;
+
+    void RestoreReadTasksFromSource(TStageInfo& stageInfo, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot, ui32 scheduledTaskCount);
+
+    void FillStages();
 
     void BuildKqpStageChannels(TStageInfo& stageInfo, ui64 txId, bool enableSpilling, bool enableShuffleElimination);
     bool IsCrossShardChannel(const NYql::NDq::TChannel& channel) const;
@@ -472,9 +481,6 @@ private:
 
     void SerializeTaskToProto(const TTask& task, NYql::NDqProto::TDqTask* result, bool serializeAsyncIoSettings) const;
 
-    std::pair<ui32, TKqpTasksGraph::TTaskType::ECreateReason> GetMaxTasksAggregation(const TStageInfo& stageInfo, const ui32 previousTasksCount, const ui32 nodesCount) const;
-    std::pair<ui32, TKqpTasksGraph::TTaskType::ECreateReason> GetScanTasksPerNode(const TStageInfo& stageInfo, const bool isOlapScan, const ui64 nodeId, bool enableShuffleElimination = false) const;
-
     void FillSecureParamsFromStage(THashMap<TString, TString>& secureParams, const NKqpProto::TKqpPhyStage& stage) const;
 
     void BuildExternalSinks(const NKqpProto::TKqpSink& sink, TKqpTasksGraph::TTaskType& task) const;
@@ -484,17 +490,16 @@ private:
     void FillKqpTableSinkSettings(NKikimrKqp::TKqpTableSinkSettings& settings, const std::vector<std::pair<ui64, i64>>& internalSinksOrder, const TKqpTasksGraph::TTaskType& task) const;
 
     std::vector<std::pair<ui64, i64>> BuildInternalSinksPriorityOrder() const;
-    TString ReplaceStructuredTokenReferences(const TString& token) const;
 
-private:
+protected:
     const TVector<IKqpGateway::TPhysicalTxData>& Transactions;
     NKikimr::NKqp::TTxAllocatorState::TPtr TxAlloc;
+    const TIntrusiveConstPtr<NACLib::TUserToken> UserToken;
+
+private:
     const NKikimrConfig::TTableServiceConfig::TAggregationConfig AggregationSettings;
     TKqpRequestCounters::TPtr Counters;
     TActorId BufferActorId; // TODO: not sure if it belongs here
-    const TIntrusiveConstPtr<NACLib::TUserToken> UserToken;
-
-    TMaxTasksGraph MaxTasksGraph;
 };
 
 } // namespace NKikimr::NKqp
