+++ ydb/core/kqp/query_data/kqp_prepared_query.h
@@ -148,6 +148,7 @@ public:
 class TPreparedQueryHolder {
 private:
     YDB_ACCESSOR_DEF(TLlvmSettings, LlvmSettings);
+    YDB_ACCESSOR(bool, UseNewKqpTasksGraph, false);
     std::shared_ptr<const NKikimrKqp::TPreparedQuery> Proto;
     std::shared_ptr<TPreparedQueryAllocHolder> Alloc;
     TVector<TString> QueryTables;
